chore: split render-handlers by domain (#split-render-handlers)
Extract 1099-line ingest/render-handlers.ts into ingest/render/*:

- shared.ts — RenderHandlersConfig, RenderResponse, ProductCountersSource
  + shared helpers (itemFromRow, notFound, buildFragmentUrl,
  isProfileRoutable, loadFeedItems, ownerStatsView,
  injectProfileSentinel)
- home.ts — / + /feed/items (+ top-today)
- gallery.ts — /gallery + /gallery/items
- drawing.ts — /d/<id> + /embed/<id> (ancestor chain cap 8)
- profile.ts — /u/<username> + /u/<username>/items
- streak.ts — /u/<username>/streak (month blocks)
- bookmarks.ts — /u/<username>/bookmarks + /me/bookmarks/feed
- follows.ts — /u/<username>/{followers,following} + thumbs
- products.ts — /products
- prompts.ts — /prompts + /prompts/<slug> + items
- design.ts — /design
- feed.ts — /feed.rss
- index.ts — barrel for new code

Keep ingest/render-handlers.ts as a backward-compatible re-export
barrel so existing imports (routes.ts, lambda.ts, dev-server.ts,
tests) keep working in this PR — no caller changes.

Net -62 lines moved, 0 behavior change. Pure file move.

Test plan:
- typecheck (tsc -b --noEmit) — pass
- build (vite build 64 modules) — pass
- test (854 tests, 122 suites) — pass
- lint (check-proxy + check-routes) — pass
- prettier --check — pass

Fixes #298 (proposal 4a).
